Update build.yml
`pull_request_target` doesn't do quite what we intended it to do, and is giving us a false sense of security over what pull requests are passing and which ones are broken.  I'm taking it off for now until I have a chance to look into this and get it working correctly for forked repositories.