This is a complete versioning system rework. (#629)
* This is a complete versioning system rework.

First, setuptools has pushed us toward using a `pyproject.toml` and `setup.cfg` in lieu of having a setuptools.setup() function call specify
the package parameters.

In addition, we have a much easier way of injecting our prerelease tags. We still technically call `setuptools.setup()`, but it knows to
rely on pyproject.toml and setup.cfg to provide all the parameters, but also to use a commandline argument for any possible version appendix to use.

However, this information is written to the sdist package, so the prerelease version numbers are maintained on publish. This is what we want.

If for some reason we are trying to get the version from source code, not the sdist package, we still try to use the `setup.cfg` file,
and we append a dev tag based on today's date and time. This is most common for the developers who work on graspologic itself.

The thing this fixes is that using `pip install -r requirements.txt` no longer changes the source tree and include a then-materialized
version number. This should avoid many of the accidental commits of `version.txt` that plagued and confused everyone in the past.

IMPORTANT NOTE: We may need to do some minor tweaks to fix the build workflow stuff because some of it isn't testable until we're on
dev or main.

Fixes #554

* This file is necessary for netlify to work. Eventually we want to remote netlify and rely on RTD, but until I get that all done this file needs to stay.

* Netlify doesn't support 3.8 apparently

* Ran black formatting

* This will at least unbreak the doc build on netlify, but it's unclear if it will give us the right level of detail re: version numbers in the generated documentation.  I'm pretty sure it won't.

* Error in workflow