Remove temporary publication from doc-split branch
The doc-split branch was added as a valid github.ref in our workflow to allow publication to gh-pages.  This was fine as we were testing things out but now that we have merged back to dev we do not want this branch to be capable of publishing documentation site from it.