Allowing files with only (source, target)
If there is no weight a weight of 1.0 will be added for each edge that
is missing a weight. A warning message will be logged that there are no
weights. If all all edges have the same weight there may be an issue
if the number of edges is greater than max_edges and we try to trim
edges.