Adds a pipeline module with adjacency spectral embedding for networkx only in it.
One major aspect of this seemingly innocuous change is that we want to be able to seed the randomized svd.  This in turn has had a cascade of
changes in most of the spectral embeddings and their tests.

It wasn't until I was fixing these tests that I found out we weren't writing unittests using the unittest module, but instead relying solely
on pytest to find and execute all of our tests.  It has done this well for the CLI, but it breaks a lot of IDE test runner code,
making it harder to debug tests since none are actually being executed in intellij/pycharm/etc.