Remove dtype requirement in vertex nomination (#865)
* Remove dtype requirement in vertex nomination

Why are we requiring that the adjacency matrix that SVN eats is type `int`? our `sbm` simulation from `models` returns a float64. Seems weird to make an sbm model, then have to cast it to a different type to do anything.

if it really matters and we care, let's just do `X = X.astype(int)` inside the code?

* fix up VN tutorial

* allow lists for y too

* make tests pass

* update mypy

* use Union operator instead

* remove typing import

* ran isort