The commit closes #912.
There are a number of changes in this commit:
- The basic build will run on any PR unless it is a change to the 2 contributor files
- The basic build will run on any push unless it is a change to the 2 contributor files or a push to dev/main
- A scheduled build will run every morning at 8am-ish pacific
- A publish build will run on any push to dev/main, provided it is not a change to the 2 contributor files

For reference, the contributor files are a repository-only change - they do not impact the build artifact at all, hence their exclusion.

Unfortunately we can't really test our builds locally, so we're going to hope this works.  Any further commits and pushes will doubtless fix any issues.