With changes in mypy and upstream typesheds, some of our downstream uses that were valid before became marked as invalid. Very few of them were actual bugs (though a few were), but regardless they were adjusted so that they did not, by and large, result in further mypy complaints.
I will note that there are still known bugs in graspologic.models.sbm_estimators, and I don't know enough about the intent
or the numpy wizardry to make any changes with confidence.  Still this gets rid of the bulk of the issues and I'll
rely on bdpedigo to shine a brilliant light on the rest.